requirements_dev.txt
PyQt5
pyqtgraph
argparse
glob2
h5py
hashlib
hdf5plugin
json
logging
python-magic
math
matplotlib
numpy
os
psutil
queue
skimage
socket
struct
sys
time
torch
traceback
unittest
uuid
zmq